requirements.txt
requests
python-dotenv
py-clob-client
web3
